Some changes for Working better on Wine (Linux)
Some commands as Window and Form now can move properly the basic form.
Also removed a page fault, at the end of running (was not every time,
and I found that some forms must not unloaded, as it did in a Windows
os, so unloandig done by WIne, else Wine try to unloiad and get page
fault).