Revision 5 (Version 8.8)
Modules can be called decorated with alter versions of internal modules.
This is not a by reference call, because modules can't be passed as
parammeters. What we can, is to prepare modules before run a module, so
in module code definitions are droped as the decor modules are locked
and can't be changed.
to call a module there are two ways. By using the name only and give a
list of parammeters (wihtout parenthesis in name), and this call can be
decorated using ";". The other call is by using Call command and that
call can't be decorated. The second one are a call like a function, so
we call recuirsve a module. So in a recuirsve call a module can't
decorate itself.