Revision 9 (Version 8.9)
some minor bugs. Added handle for this type of reference:
m(10)(1)(2,3)(2)(0)("alfa")(100)
where m is an invenrory (a pair of keys and values), (!) is for array,
(2,3) is parameter in a lambda function, (2) is a parameter in a lambda
function returned from previouw one, and (0) is an index for array
returned from previous lambda function, which array item is an
inventory, where "alfa" is the key and return a group which have a value
but need a parameter so we place 100.

dim a(10)
a(1)=lambda -> {
read x, y
=lambda x, y (b) -> {
group z {
k=lambda ->500

x=10
Value (x) {
=2000*x
}
}
dim b(1)
inventory kk = "alfa":=group(z)
b(0)=kk
=b()
}
}
inventory m=10:=a()
Print m(10)(1)(2,3)(2)(0)("alfa")(100)