Revision 23 (Version 8.9)
Found a hang situtation when we edit text in console, using screen
editor, and we have onpe a user form. This happen because of subbclassig
for middle button in mouse. Now works nice (now every time we loose
focus from console form (form1) we did an unhook, and next time we get
focus and if needed we get a hook to mouse event.