Revision 24 (Version 8.9)
Now we can pass reference (a special hard link) from a property, wich
need a parameter, in an object (com type) to an item in array or
iventory list. So if control is an inventory and Enabled() is a property
with a parameter, we can Append control, key1:=Enabled() and we can
access this using:
control(key)(2)
(where 2 is value for parameter) for write or read value. Also if we
have a List$() we can do the same as for Enabled() and we can use:
control$(key)(2)
For arrays we can use a refernce to array, and do an Append using auto
array format (,),  like:
Control=(,)   ' an array with no length, a zero one
Append control, (Enabled(),)

for reading or write value