Rev 6 (Version 9.0)
Ternary operator (as functions If() and If$()). Only selected exrpession
evaluated. Elvis operator work too as variant of If(). New spaceship
operator <=> for expressions, is like Compare(), but the last one is for
variables/arrayitems only,
Also in if then , after Else we can write a number to jump, wiithout
goto (but goto needed if we use a label with letters). In M2000  we can
use numbers in lines with commands, but only comments in lines with
labels as words.