Revision 30  Version 9.0
Some minor fixes, plus a change in Load, so we can load library of M2000
functions/modules, and can contain commands, but every call to module
skiped. So we may have library and program in one file, and we can call
it using USE (start a new M2000.exe)  or we can load it using Load.