Revision 26 Version 9.4
Fix the Let statement when we assign values to array items. A bug in
later revisions prevents array items to get objects (although without
using Let we get this functionality). A Let statement first execute the
right expression and then the left expression (and the expression for
each array dimension).
Also new we can use Let A(1)="string value" which we can do that using
A(1):="string value", but now using the right evalution first rule.