Version 9.5
A lot of improvements. Now array items and static variables can be
passed by reference (using a copy in copy out mechanism).
Filter/Map/Fold plus some more for arrays (we can use them using a #
before, so A#sum() return the sum of all items of A (a pointet to array,
or tuple).
New function StrRev$() for string reverse.